Update statistics-resampling.yaml
Version 5.5.4 sees a move of the manual to the `/docs` folder and it's publication online at GitHub pages. Some minor improvements have been made to the installer and mex files. A Matlab toolbox has also installer has also been created and is available in `/matlab`. Demos have been added to `bootlm` to illustrate interpretation of ANOVA on datasets with unbalanced sample sizes and type I sums-of-squares. It is also illustrated how `bootlm` can be used for ANOVA with type II sums-of-squares in ANOVA and ANCOVA models of 2 predictors without an interaction. A demo has also been added to `bootci` and `bootknife` to illustrate how it can be used for obtaining confidence intervals for logistic regression coefficients.