Remove packajoozle, bc it is obsolete and/or has quality problems (#544)
Packajoozle was never ready for prime time - never had an official release - and hasn't had an update in 5 years. So it has probably been obsoleted and/or broken by the changes to pkg() and related Octave packaging stuff in the last few releases, and it wasn't all that good to begin with. I wouldn't recommend anyone use packajoozle at this point, and it's probably best to not include it in public indexes like this.

See discussion in https://octave.discourse.group/t/suggestions-for-clearer-name-besides-forge-to-use-with-pkg-command/6358/51?u=apjanke and following comments.