CI: Demote error to warning if sha256 checksum in .yaml is empty.
There can be legit reasons why the sha256 checksum of a version can be
empty. E.g., that version could track a "rolling release" of a package.

In most cases, an empty sha256 checksum is probably an oversight that
should be fixed. So, emit a warning if it is empty, but continue testing
the package.